Move sales tests to a new file
